WW-5459 Moves ActionChainResult into org.apache.struts2.result package
