+++ core/src/main/java/org/apache/struts2/result/Result.java
@@ -18,10 +18,8 @@
  */
 package org.apache.struts2.result;
 
-import org.apache.struts2.Action;
-import org.apache.struts2.ActionInvocation;
-
 import org.apache.struts2.action.Action;
+import org.apache.struts2.ActionInvocation;
 
 import java.io.Serializable;
