+++ core/src/test/java/org/apache/struts2/ActionChainResultTest.java
@@ -18,6 +18,7 @@
  */
 package org.apache.struts2;
 
+import org.apache.struts2.result.ActionChainResult;
 import org.apache.struts2.result.Result;
 import org.junit.Test;
