+++ core/src/test/java/org/apache/struts2/config/providers/ConfigurationProviderOgnlAllowlistTest.java
@@ -51,7 +51,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("org.apache.struts2.mock.MockResult"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.ActionSupport"),
-                Class.forName("org.apache.struts2.ActionChainResult"),
+                Class.forName("org.apache.struts2.result.ActionChainResult"),
                 Class.forName("org.apache.struts2.TextProvider"),
                 Class.forName("org.apache.struts2.interceptor.NoOpInterceptor"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
@@ -100,7 +100,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("java.io.Serializable"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
                 Class.forName("org.apache.struts2.ActionSupport"),
-                Class.forName("org.apache.struts2.ActionChainResult"),
+                Class.forName("org.apache.struts2.result.ActionChainResult"),
                 Class.forName("org.apache.struts2.TextProvider"),
                 Class.forName("org.apache.struts2.interceptor.NoOpInterceptor"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
