+++ core/src/test/resources/org/apache/struts2/config/providers/xwork-include-parent.xml
@@ -27,7 +27,7 @@
 
     <package name="default">
         <result-types>
-            <result-type name="chain" class="org.apache.struts2.ActionChainResult" default="true"/>
+            <result-type name="chain" class="org.apache.struts2.result.ActionChainResult" default="true"/>
         </result-types>
         <interceptors>
             <interceptor name="params" class="org.apache.struts2.interceptor.parameter.ParametersInterceptor"/>
