+++ core/src/test/resources/org/apache/struts2/config/providers/xwork-test-results.xml
@@ -26,7 +26,7 @@
 	<include file="xwork-test-beans.xml" />
     <package name="default">
         <result-types>
-            <result-type name="chain" class="org.apache.struts2.ActionChainResult" default="true"/>
+            <result-type name="chain" class="org.apache.struts2.result.ActionChainResult" default="true"/>
             <result-type name="mock" class="org.apache.struts2.mock.MockResult"/>
         </result-types>
