+++ core/src/test/resources/xwork-test-validation.xml
@@ -26,7 +26,7 @@
 	<include file="xwork-test-beans.xml" />
     <package name="xwork-test-default">
         <result-types>
-            <result-type name="chain" class="org.apache.struts2.ActionChainResult" default="true"/>
+            <result-type name="chain" class="org.apache.struts2.result.ActionChainResult" default="true"/>
             <result-type name="void" class="org.apache.struts2.VoidResult"/>
             <result-type name="mock" class="org.apache.struts2.mock.MockResult"/>
         </result-types>
