+++ plugins/spring/src/test/java/org/apache/struts2/spring/SpringObjectFactoryTest.java
@@ -19,7 +19,7 @@
 package org.apache.struts2.spring;
 
 import org.apache.struts2.action.Action;
-import org.apache.struts2.ActionChainResult;
+import org.apache.struts2.result.ActionChainResult;
 import org.apache.struts2.ActionContext;
 import org.apache.struts2.ModelDrivenAction;
 import org.apache.struts2.ObjectFactory;
